+++ mod-oai-pmh/src/test/resources/firebird/oaipmh/oaipmh-getrecord-deleted-C729193.feature
@@ -1,4 +1,5 @@
 @parallel=false
+@C729193
 Feature: GetRecord: SRS & Inventory - Verify that set for deletion MARC Instances are harvested
 
   # TestRail Case ID: C729193
