+++ mod-oai-pmh/src/test/resources/firebird/oaipmh/oaipmh-getrecord-deleted-srs-leader-past-C375977.feature
@@ -1,4 +1,5 @@
 @parallel=false
+@C375977
 Feature: GetRecord: SRS & Inventory - Verify SRS LDR/05=d records are harvested as deleted
 
   # JIRA: MODOAIPMH-613, MODOAIPMH-614
